working step1, changed bathrooms to float

+++ data/request/createRentalRequest.go
@@ -11,7 +11,7 @@ type CreateRentalRequest struct {
 	LocationID  uint    `json: validate:"required"`
 	AccountID   uint    `json: validate:"required"`
 	Bedrooms    uint    `json: validate:"required"`
-	Bathrooms   uint    `json: validate:"required"`
+	Bathrooms   float64 `json: validate:"required"`
 	NightlyRate float64 `json: validate:"required"`
 	Description string  `json: validate:"required"`
 	Thumbnail   string
